+++ components/library/src/content/section.rs
@@ -13,6 +13,7 @@ use utils::site::get_reading_analytics;
 use utils::templates::render_template;
 
 use content::file_info::FileInfo;
+use content::has_anchor;
 use content::ser::SerializingSection;
 use library::Library;
 
@@ -56,6 +57,11 @@ pub struct Section {
     pub lang: String,
     /// Contains all the translated version of that section
     pub translations: Vec<Key>,
+    /// Contains the internal links that have an anchor: we can only check the anchor
+    /// after all pages have been built and their ToC compiled. The page itself should exist otherwise
+    /// it would have errored before getting there
+    /// (path to markdown, anchor value)
+    pub internal_links_with_anchors: Vec<(String, String)>,
     /// Contains the external links that need to be checked
     pub external_links: Vec<String>,
 }
@@ -87,6 +93,7 @@ impl Section {
             reading_time: None,
             lang: String::new(),
             translations: Vec::new(),
+            internal_links_with_anchors: Vec::new(),
             external_links: Vec::new(),
         }
     }
@@ -193,6 +200,7 @@ impl Section {
         self.content = res.body;
         self.toc = res.toc;
         self.external_links = res.external_links;
+        self.internal_links_with_anchors = res.internal_links_with_anchors;
 
         Ok(())
     }
@@ -229,6 +237,10 @@ impl Section {
             .collect()
     }
 
+    pub fn has_anchor(&self, anchor: &str) -> bool {
+        has_anchor(&self.toc, anchor)
+    }
+
     pub fn to_serialized<'a>(&'a self, library: &'a Library) -> SerializingSection<'a> {
         SerializingSection::from_section(self, library)
     }
@@ -260,6 +272,7 @@ impl Default for Section {
             word_count: None,
             lang: String::new(),
             translations: Vec::new(),
+            internal_links_with_anchors: Vec::new(),
             external_links: Vec::new(),
         }
     }