+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix image processing not happening if called from the template
 - Add a `zola check` command to that validates the site and checks all external links
 - Sections can have `aliases` as well
+- Anchors in internal links are now checked for existence
 
 ## 0.7.0 (2019-04-28)